Made some changes to the comments and started implementing Tiptap Rich text editor.

+++ app/api/posts/[id]/route.js
@@ -1,6 +1,7 @@
 import { NextResponse } from "next/server";
 import { getPostById, updatePost, deletePost } from "@/queries/posts";
 import dbConnect from "@/lib/mongo";
+import { auth } from "@/auth";
 
 
 export const GET = async (request, { params }) => {
@@ -24,49 +25,56 @@ export const GET = async (request, { params }) => {
     }
 };
 
-export const PATCH = async (request, { params }) => {
-    const { title, body } = await request.json()
+export const PATCH = auth(async function PATCH(request, { params }) {
 
+    if (request.auth.user.role === 'admin') {
+        const { title, body } = await request.json()
+        try {
+            await dbConnect();
+            console.log("Database connected");
+            console.log(params.id)
+            const id = params.id
 
-    try {
-        await dbConnect();
-        console.log("Database connected");
-        console.log(params.id)
-        const id = params.id
+            const post = await updatePost(id, { title, body });
+            console.log("Updated Post:", post);
 
-        const post = await updatePost(id, { title, body });
-        console.log("Updated Post:", post);
-
-        return new NextResponse(JSON.stringify(post), {
-            status: 200
-        });
-    } catch (error) {
-        console.error("Error updating post:", error);
-        return new NextResponse(error.message, {
-            status: 500
-        });
+            return new NextResponse(JSON.stringify(post), {
+                status: 200
+            });
+        } catch (error) {
+            console.error("Error updating post:", error);
+            return new NextResponse(error.message, {
+                status: 500
+            });
+        }
+    } else {
+        return NextResponse.json({ message: "Not authorized" }, { status: 401 })
     }
-};
+});
 
 
-export const DELETE = async (request, { params }) => {
+export const DELETE = auth(async function DELETE(request, { params }) {
 
-    try {
-        await dbConnect();
-        console.log("Database connected");
-        console.log(params.id)
-        const id = params.id
+    if (request.auth.user.role === 'admin') {
+        try {
+            await dbConnect();
+            console.log("Database connected");
+            console.log(params.id)
+            const id = params.id
 
-        const post = await deletePost(id);
-        console.log("Deleted Post:", post);
+            const post = await deletePost(id);
+            console.log("Deleted Post:", post);
 
-        return new NextResponse(JSON.stringify(post), {
-            status: 200
-        });
-    } catch (error) {
-        console.error("Error Deleting post:", error);
-        return new NextResponse(error.message, {
-            status: 500
-        });
+            return new NextResponse(JSON.stringify(post), {
+                status: 200
+            });
+        } catch (error) {
+            console.error("Error Deleting post:", error);
+            return new NextResponse(error.message, {
+                status: 500
+            });
+        }
+    } else {
+        return NextResponse.json({ message: "Not authorized" }, { status: 401 })
     }
-};
+});
